Bootstrap the Cosmos DB ##111
This patch adds a method for making bundle specs in the state store during startup.
It creates the container and items once and then checks if they exist and exits if they do.

Eventually we should add an API which makes it possible to register bundle specs during authoring experience