Airlock processor - function app based - First commit
Adding the Airlock processor, a function app based orchestrator for the airlock component.
In this commit, the AL processor is wired to a SB queue as input and EG topic as output, gets a message and sends one back.
This commit contains all the IaC changes to deploy this new component